requirements.txt
requests==2.24.0
APScheduler==3.6.3
sseclient-py==1.7